requirements.txt
Django==1.7
South==1.0
argparse==1.2.1
distribute==0.6.24
dj-database-url==0.3.0
djangorestframework==2.4.2
psycopg2==2.5.4
wsgiref==0.1.2
